requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
google-genai>=1.0.0
python-dotenv>=1.0.0
openai>=1.0.0
requests>=2.28.0
eventregistry>=9.0
tavily-python>=0.3.3
fastapi>=0.100.0
uvicorn>=0.20.0
PyPDF2>=3.0.0
